float -> double

+++ include/GUI.hpp
@@ -16,11 +16,11 @@ public:
   GUIContainer *get_root_container() const;
   void add_container(GUIContainer *container);
   void add_text(GUIText *text);
-  void on_window_resize(float width, float height);
-  void on_mouse_move(float x1, float y1, float x2, float y2);
-  void on_mouse_down(float x, float y);
-  void on_mouse_up(float x, float y);
-  void on_mouse_drag(float x1, float y1, float x2, float y2);
+  void on_window_resize(double width, double height);
+  void on_mouse_move(double x1, double y1, double x2, double y2);
+  void on_mouse_down(double x, double y);
+  void on_mouse_up(double x, double y);
+  void on_mouse_drag(double x1, double y1, double x2, double y2);
   void reset_focus();
   void render() const;
   ~GUI();
