+++ include/GUI/Layout.hpp
@@ -19,9 +19,9 @@ public:
   GUILayout(GUIUnit width, GUIUnit height, GUISegment left, GUISegment top);
   GUILayout(GUISegment left, GUISegment top, GUISegment right, GUISegment bottom);
   GUILayout(GUISegment left, GUISegment top);
-  float get_left();
-  void move(float dx, float dy);
-  SDL_FRect calculate(float parent_width, float parent_height);
+  double get_left();
+  void move(double dx, double dy);
+  SDL_FRect calculate(double parent_width, double parent_height);
 };
 
 #endif
